Rename package
Rename package from `lab` to `x-date-pickers` to reflect MUI changes
