+++ packages/formik-mui-x-date-pickers/README.md
@@ -1,7 +1,7 @@
 # Formik MUI
 
 ![](https://github.com/stackworx/formik-mui/workflows/Build%20formik-mui/badge.svg)
-![](https://github.com/stackworx/formik-mui/workflows/Build%20formik-mui-lab/badge.svg)[![license](https://badgen.now.sh/badge/license/MIT)](./LICENSE)
+![](https://github.com/stackworx/formik-mui/workflows/Build%20formik-mui-x-date-pickers/badge.svg)[![license](https://badgen.now.sh/badge/license/MIT)](./LICENSE)
 
 Bindings for using [Formik](https://github.com/jaredpalmer/formik) with [MUI](https://mui.com/).
