+++ packages/formik-mui-x-date-pickers/src/StaticDatePicker.tsx
