+++ packages/formik-mui-x-date-pickers/src/StaticDateTimePicker.tsx
