+++ packages/formik-mui-x-date-pickers/src/__tests__/StaticDatePicker.test.tsx
