+++ packages/formik-mui-x-date-pickers/src/__tests__/StaticDateTimePicker.test.tsx
