+++ stories/DatePicker.story.tsx
@@ -5,10 +5,10 @@ import { subDays } from 'date-fns';
 import { Field, Form, Formik } from 'formik';
 import * as React from 'react';
 import * as yup from 'yup';
-import { DatePicker } from '../packages/formik-mui-lab/src/DatePicker';
-import { DesktopDatePicker } from '../packages/formik-mui-lab/src/DesktopDatePicker';
-import { MobileDatePicker } from '../packages/formik-mui-lab/src/MobileDatePicker';
-import { StaticDatePicker } from '../packages/formik-mui-lab/src/StaticDatePicker';
+import { DatePicker } from '../packages/formik-mui-x-date-pickers/src/DatePicker';
+import { DesktopDatePicker } from '../packages/formik-mui-x-date-pickers/src/DesktopDatePicker';
+import { MobileDatePicker } from '../packages/formik-mui-x-date-pickers/src/MobileDatePicker';
+import { StaticDatePicker } from '../packages/formik-mui-x-date-pickers/src/StaticDatePicker';
 import FormValues from './FormValues';
 import Wrapper from './Wrapper';
