+++ stories/TimePicker.story.tsx
@@ -4,10 +4,10 @@ import { action } from '@storybook/addon-actions';
 import { Field, Form, Formik } from 'formik';
 import * as React from 'react';
 import * as yup from 'yup';
-import { DesktopTimePicker } from '../packages/formik-mui-lab/src/DesktopTimePicker';
-import { MobileTimePicker } from '../packages/formik-mui-lab/src/MobileTimePicker';
-import { StaticTimePicker } from '../packages/formik-mui-lab/src/StaticTimePicker';
-import { TimePicker } from '../packages/formik-mui-lab/src/TimePicker';
+import { DesktopTimePicker } from '../packages/formik-mui-x-date-pickers/src/DesktopTimePicker';
+import { MobileTimePicker } from '../packages/formik-mui-x-date-pickers/src/MobileTimePicker';
+import { StaticTimePicker } from '../packages/formik-mui-x-date-pickers/src/StaticTimePicker';
+import { TimePicker } from '../packages/formik-mui-x-date-pickers/src/TimePicker';
 import FormValues from './FormValues';
 import Wrapper from './Wrapper';
